feat: enable many group IDs per invite & use query instead of aggregate
