+++ src/resolvers/Invitation/index.js
@@ -1,11 +1,10 @@
 import resolveShopFromShopId from "@reactioncommerce/api-utils/graphql/resolveShopFromShopId.js";
 import { encodeInvitationOpaqueId } from "../../xforms/id.js";
-import groups from "./groups.js";
 import invitedBy from "./invitedBy.js";
 
 export default {
   _id: (node) => encodeInvitationOpaqueId(node._id),
-  groups,
+  groups: (parent, _, context) => context.queries.groupsById(context, parent.groupIds || [parent.groupId]),
   invitedBy,
   shop: resolveShopFromShopId
 };
