+++ src/schemas/inviteShopMember.graphql
@@ -91,6 +91,12 @@ extend type Query {
     "The shop IDs to get invitations for"
     shopIds: [ID],
 
+    "Return only results that come after this cursor. Use this with `first` to specify the number of results to return."
+    after: ConnectionCursor,
+
+    "Return only results that come before this cursor. Use this with `last` to specify the number of results to return."
+    before: ConnectionCursor,
+
     "Return at most this many results. This parameter may be used with either `after` or `offset` parameters."
     first: ConnectionLimitInt,
